+++ src/main/scala/info/gratour/common/service/ShutdownRegistry.scala
@@ -1,8 +1,9 @@
 package info.gratour.common.service
 
 import java.io.Closeable
-
 import com.typesafe.scalalogging.Logger
+
+import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
 
 object ShutdownRegistry {
 
@@ -17,12 +18,42 @@ object ShutdownRegistry {
       }
     }
   }
+  private val startup: AtomicBoolean = new AtomicBoolean()
+  private val delayedShutdown: AtomicBoolean = new AtomicBoolean()
+  private val shutdownExecuted: AtomicBoolean = new AtomicBoolean()
+
   val list: java.util.List[Entry] = new java.util.ArrayList[Entry]()
 
+  def beginStartup(): Unit = {
+    startup.set(true)
+  }
+
+  def endStartup(): Unit = {
+    startup.set(false)
+    if (delayedShutdown.get()) {
+      logger.debug("Execute delayed shutdown.")
+      shutdown()
+    }
+  }
+
   def shutdown(): Unit = {
-    this.synchronized {
-      list.forEach { e =>
-        e.close()
+    if (startup.get()) {
+      logger.debug("shutdown() called in startup processing, delay it.", new Throwable)
+      delayedShutdown.set(true)
+      return
+    }
+
+    if (shutdownExecuted.compareAndSet(false, true)) {
+
+      this.synchronized {
+        logger.debug("Starting shutdown() processing, registry: ")
+        list.forEach(e => {
+          logger.debug("  " + e.name)
+        })
+
+        list.forEach { e =>
+          e.close()
+        }
       }
     }
   }
@@ -37,6 +68,35 @@ object ShutdownRegistry {
     this.synchronized {
       list.add(Entry(name, autoCloseable))
     }
+  }
+
+  private class ShutdownRunnable(delaySeconds: Int) extends Runnable {
+    override def run(): Unit = {
+      if (delaySeconds > 0)
+        Thread.sleep(delaySeconds * 1000)
+
+      shutdown()
+    }
+  }
+
+  def scheduleShutdown(delaySeconds: Int): Unit = {
+    val t = new Thread(new ShutdownRunnable(delaySeconds))
+    t.start()
+  }
+
+  private class DelayedRunnable(delaySeconds: Int, runnable: Runnable) extends Runnable {
+    override def run(): Unit = {
+
+      if (delaySeconds > 0)
+        Thread.sleep(delaySeconds * 1000)
+
+      runnable.run()
+    }
+  }
+
+  def delayRun(delaySeconds: Int, runnable: Runnable): Unit = {
+    val t = new Thread(new DelayedRunnable(delaySeconds, runnable))
+    t.start()
   }
 
 }